backend/requirements.txt
fastapi
uvicorn
pandas
numpy
sqlalchemy
pymysql
python-multipart
openai
python-dotenv
